mcp-gcp-cloud/requirements.txt
mcp>=1.0.0
fastmcp>=0.9.0
google-cloud-compute>=1.19.0
google-cloud-storage>=2.17.0
google-cloud-functions>=1.16.0
google-cloud-monitoring>=2.21.0
google-cloud-billing>=1.12.0
google-auth>=2.29.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
